@@ -3,16 +3,17 @@ requires = ['poetry-core']
 build-backend = 'poetry.core.masonry.api'
 
 [tool.poetry]
-name = 'convert_redirects'
+name = 'create_redirects'
 version = '0.0.1.dev'
 description = ''
 authors = ['Andrew Aldridge <i80and@foxquill.com>']
 
 [tool.poetry.dependencies]
-python = '>=3.8,<4.0'
+python = '>=3.9,<4.0'
 boto3 = '1.34.144'
 requests = '2.32.3'
 pyyaml = "^6.0.2"
+pandas = "^2.2.3"
 
 [tool.poetry.group.dev.dependencies]
 mypy = "^1.10.0"
